requirements.txt
click==8.1.8
natsort==8.4.0
numpy==2.1.1
pandas==2.2.3
Pillow==10.4.0
Requests==2.32.3
pypdfium2==4.30.0
split-image==2.0.1